+++ lists/mirror-tv/Sponsor.js
@@ -1,0 +1,53 @@
+const { Integer, Relationship } = require('@keystonejs/fields')
+
+const { byTracking } = require('@keystonejs/list-plugins')
+const { atTracking } = require('../../helpers/list-plugins')
+const {
+    admin,
+    moderator,
+    editor,
+
+    contributor,
+    allowRoles,
+} = require('../../helpers/access/mirror-tv')
+const ImageRelationship = require('../../fields/ImageRelationship')
+const cacheHint = require('../../helpers/cacheHint')
+
+module.exports = {
+    fields: {
+        sortOrder: {
+            label: '排序順位',
+            type: Integer,
+            isUnique: true,
+        },
+        topic: {
+            label: '專題',
+            type: Relationship,
+            ref: 'Topic',
+            isRequired: true,
+        },
+        logo: {
+            label: '首圖',
+            type: ImageRelationship,
+            ref: 'Image',
+        },
+    },
+    plugins: [
+        atTracking({
+            hasNowBtn: false,
+            isReadOnly: true,
+        }),
+        byTracking(),
+    ],
+    access: {
+        update: allowRoles(admin, moderator, editor, contributor),
+        create: allowRoles(admin, moderator, editor, contributor),
+        delete: allowRoles(admin, moderator),
+    },
+    adminConfig: {
+        defaultColumns: 'sortOrder, topic',
+        defaultSort: '-sortOrder',
+    },
+    labelField: 'sortOrder',
+    cacheHint: cacheHint,
+}